+++ client/src/pages/MainPage.js
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 import Carousel from '../components/Carousel';
 import QuestionCards from '../components/QuestionCards';
 import Tagbox from '../components/Tagbox';
-
 const Main = styled.main`
   max-width: 1200px;
   width: 100%;
